refactor: widen SP-offset chain to i128; delete int_const_i64
SpExpr.offset was i64, forcing every SP-relative constant addend through the
narrowing int_const_i64 reader. Per request, widen the whole offset/size/address
arithmetic chain to i128 so the SP reasoning never narrows, and delete
int_const_i64 (its only prod caller was the SP analyzer):

- strider-ir: Function::stack_offsets side-table + set/get_stack_offset +
  compact remap, FunctionArgSource::Stack{offset}, dot display -> i128;
  int_const_i64 deleted from IRViewer (int_const_i128 is the signed SSoT).
- strider-opt sp_expr: SpExpr.offset/shifted, AddrClass::{SpRooted,Constant},
  offset_range_verdict/alias_verdict sizes, ranges_disjoint (+ i64::MAX ->
  i128::MAX sentinel), cfg load sizes; analyzer reads addends via
  int_const_i128. eval.rs Abs::SpRel offset likewise. Also adopt
  single_value_output in load_class_and_size.
- strider-target: StackArgs.base_offset/increment + offset_of/index_of/slot_of/
  slots_spanned -> i128 (the positional-arg offsets feed SpExpr).
- strider-pattern: function_arg_stack / Load/Store stack_offset filters -> i128.
- strider-py: function_arg_stack / source_stack / CC stack-arg params -> i128
  (Python ints are arbitrary-precision, so the binding is unaffected).

Pure type widening; no alias-verdict / pass-ordering / overflow-fail-closed
logic changed. CC boundary-overflow tests recomputed for the wider type (those
inputs no longer overflow at i128, so they scale exactly instead of saturating;
the saturation guards remain). Gate: cargo test --workspace exit 0 + clippy
--all-targets clean + pytest 870.

+++ crates/strider-ir/src/function/data.rs
@@ -226,7 +226,7 @@ pub struct Function {
     /// Populated by the `StackOffsetDetect` classifier.  The phi-of-offsets
     /// case (address is a phi of different constants per branch) is not
     /// recorded — consumers can re-decompose via `decompose_sp` if needed.
-    stack_offsets: SecondaryMap<NodeId, Option<(ValueId, i64)>>,
+    stack_offsets: SecondaryMap<NodeId, Option<(ValueId, i128)>>,
 
     /// O(1) varnode → `InitialVar(vn)` node-id accelerator for
     /// indirect-resolve sites and the lifter's lazy `read_or_init_var`
@@ -765,13 +765,13 @@ impl Function {
     /// relative to; the offset is only comparable against another access's
     /// offset when their bases match.
     #[inline]
-    pub fn stack_offset(&self, id: NodeId) -> Option<(ValueId, i64)> {
+    pub fn stack_offset(&self, id: NodeId) -> Option<(ValueId, i128)> {
         self.stack_offsets[id]
     }
 
     /// Records a concrete stack slot `(base, offset)` for a Store/Load node.
     #[inline]
-    pub fn set_stack_offset(&mut self, id: NodeId, base: ValueId, offset: i64) {
+    pub fn set_stack_offset(&mut self, id: NodeId, base: ValueId, offset: i128) {
         self.stack_offsets[id] = Some((base, offset));
     }
 
@@ -976,7 +976,7 @@ impl Function {
         // translation table.  An entry whose node or base didn't survive
         // compaction is dropped (the slot becomes "unknown", which is safe —
         // consumers treat a missing entry as non-stack).
-        let mut new_stack_offsets: SecondaryMap<NodeId, Option<(ValueId, i64)>> =
+        let mut new_stack_offsets: SecondaryMap<NodeId, Option<(ValueId, i128)>> =
             SecondaryMap::new();
         for (old_id, slot) in self.stack_offsets.iter() {
             let Some((old_base, off)) = *slot else {
