+++ crates/strider-opt/src/sp_expr/ranges.rs
@@ -33,18 +33,18 @@ pub(crate) fn high_low_shift_bits(
 /// disjoint.
 ///
 /// Endpoint computations use `saturating_add` so that callers passing
-/// `size = i64::MAX` as a soundness-pessimistic fallback (e.g. when a Store's
+/// `size = i128::MAX` as a soundness-pessimistic fallback (e.g. when a Store's
 /// `value_byte_size` is unknown) cannot panic in debug or wrap in release.
 /// A saturated upper endpoint additionally short-circuits to "not disjoint"
 /// — i.e. an unknown-extent range is treated as effectively infinite in both
 /// directions, matching the conservative verdict callers expect.
 #[inline]
-pub fn ranges_disjoint(a_off: i64, a_size: i64, b_off: i64, b_size: i64) -> bool {
+pub fn ranges_disjoint(a_off: i128, a_size: i128, b_off: i128, b_size: i128) -> bool {
     let a_end = a_off.saturating_add(a_size);
     let b_end = b_off.saturating_add(b_size);
     // If either endpoint saturated, treat the corresponding range as
     // unbounded and report "not disjoint" — the conservative answer.
-    if a_end == i64::MAX || b_end == i64::MAX {
+    if a_end == i128::MAX || b_end == i128::MAX {
         return false;
     }
     a_end <= b_off || b_end <= a_off
@@ -56,11 +56,11 @@ pub fn ranges_disjoint(a_off: i64, a_size: i64, b_off: i64, b_size: i64) -> bool
 /// non-value here means malformed IR and panics rather than silently
 /// degrading the alias verdict.
 #[inline]
-pub(crate) fn store_value_byte_size(g: &Graph, store_data: ValueId) -> i64 {
+pub(crate) fn store_value_byte_size(g: &Graph, store_data: ValueId) -> i128 {
     g.value_kind(store_data)
         .as_value()
         .expect("Store data input is a value")
-        .byte_size() as i64
+        .byte_size() as i128
 }
 
 #[cfg(test)]
@@ -83,30 +83,30 @@ mod tests {
     fn ranges_disjoint_max_size_left_does_not_panic_and_is_conservative() {
         // The three memory-chain walkers (CallStackArgCollect,
         // load_forward::probe, function_args::mem_chain_is_dirty)
-        // pass `i64::MAX` as a soundness-pessimistic fallback when a Store's
-        // `value_byte_size` is unknown. With plain `+`, `a_off + i64::MAX`
+        // pass `i128::MAX` as a soundness-pessimistic fallback when a Store's
+        // `value_byte_size` is unknown. With plain `+`, `a_off + i128::MAX`
         // would panic in debug and wrap in release for any positive `a_off`.
         // ranges_disjoint must saturate cleanly and report "not disjoint"
         // (false) for any reachable load offset — the conservative verdict
         // callers depend on. SP-relative offsets in practice are small (kB
         // range), so we cover zero, modestly-negative, and modestly-positive
         // a_off values.
-        assert!(!ranges_disjoint(0, i64::MAX, 100, 4));
-        assert!(!ranges_disjoint(-1000, i64::MAX, 100, 4));
-        assert!(!ranges_disjoint(1_000_000, i64::MAX, -1_000_000, 4));
-        // Even very large positive a_off (where `a_off + i64::MAX` would
+        assert!(!ranges_disjoint(0, i128::MAX, 100, 4));
+        assert!(!ranges_disjoint(-1000, i128::MAX, 100, 4));
+        assert!(!ranges_disjoint(1_000_000, i128::MAX, -1_000_000, 4));
+        // Even very large positive a_off (where `a_off + i128::MAX` would
         // overflow without saturation) must not panic and must report
         // "not disjoint".
-        assert!(!ranges_disjoint(1, i64::MAX, 0, 4));
+        assert!(!ranges_disjoint(1, i128::MAX, 0, 4));
     }
 
     #[test]
     fn ranges_disjoint_max_size_right_does_not_panic_and_is_conservative() {
-        // Symmetric: i64::MAX on the b-side must also saturate and report
+        // Symmetric: i128::MAX on the b-side must also saturate and report
         // "not disjoint" without panicking.
-        assert!(!ranges_disjoint(100, 4, 0, i64::MAX));
-        assert!(!ranges_disjoint(100, 4, -1000, i64::MAX));
-        assert!(!ranges_disjoint(-1_000_000, 4, 1_000_000, i64::MAX));
-        assert!(!ranges_disjoint(0, 4, 1, i64::MAX));
+        assert!(!ranges_disjoint(100, 4, 0, i128::MAX));
+        assert!(!ranges_disjoint(100, 4, -1000, i128::MAX));
+        assert!(!ranges_disjoint(-1_000_000, 4, 1_000_000, i128::MAX));
+        assert!(!ranges_disjoint(0, 4, 1, i128::MAX));
     }
 }
